+++ pyproject.toml
@@ -20,9 +20,10 @@ dependencies = [
     "mlcroissant>=1.0.17",
     "numpy>=2.1.3",
     "opencv-python>=4.11.0.86",
+    "opencv-python-headless>=4.11.0.86",
     "pygame>=2.6.1",
     "pynput>=1.8.1",
-    "pyrealsense2>=2.55.1.6486",
+    # "pyrealsense2>=2.55.1.6486",
     "pyspacemouse>=1.1.4",
     "rich>=14.0.0",
     "tensorflow>=2.19.0",
