+++ scripts/mano/vitpose_model.py
@@ -1,0 +1,1 @@
+../../../hamer/vitpose_model.py
